# docs/swagger.yaml
openapi: 3.0.2
info:
  title: New API
  version: 1.0.0
  description: '### This is a Swagger UI for our games API'

paths:
  /games:
    get:
      tags:
        - games
      summary: Returns a list of all games
      responses:
        '200':
          description: List of all games
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

    post:
      tags:
        - new-game
      summary: Creates a new game
      description: Adds a new game to the list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/new-game'
      responses:
        '201':
          description: Game successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/new-game'
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

  /games/{id}:
    get:
      tags:
        - details
      summary: Returns full details of a given game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  price:
                    type: number
                    format: float
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'

components:
  schemas:
    error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Game not found"

    new-game:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
        price:
          type: number
          format: float